+++ test/3_eventemitter_websockets_embedded_sanity.js
@@ -40,11 +40,17 @@ describe('3_eventemitter_websockets_embedded_sanity', function () {
 
   after(function (done) {
 
-    publisherclient.disconnect({timeout:2000})
-      .then(listenerclient.disconnect({timeout:2000}))
-      .then(happnInstance.stop())
-      .then(done)
-      .catch(done);
+    publisherclient.disconnect({timeout:2000}, function(e){
+
+      if (e) console.warn('failed diconnecting test client');
+
+      listenerclient.disconnect({timeout:2000}, function(e){
+
+        if (e) console.warn('failed diconnecting test client');
+
+        happnInstance.stop(done);
+      })
+    });
   });
 
   var publisherclient;