+++ test/e3_client_connection_events.js
@@ -1,14 +1,11 @@
 var path = require('path');
 var filename = path.basename(__filename);
 
-var benchmarket = //require('benchmarket');
 var Promise = require('bluebird');
 var expect = require('expect.js');
 var Happn = require('../');
 
 describe(filename, function() {
-
-  benchmarket.start();
 
   var server;
 
@@ -210,8 +207,5 @@ describe(filename, function() {
       .catch(done);
 
   });
-
-  after(benchmarket.store());
-  benchmarket.stop();
 
 });