+++ test/e7_connection_conditions.js
@@ -1,8 +1,5 @@
 describe(require('path').basename(__filename), function () {
-
-  //require('benchmarket').start();
-  //after(//require('benchmarket').store());
-
+  
   var expect = require('expect.js');
   var happn = require('../lib/index');
   var service = happn.service;